reorganize tests and add test for writableAsyncBeacon.tryCatch
