+++ state_beacon/test/src/async_value_test.dart
@@ -1,14 +1,14 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:state_beacon/state_beacon.dart';
 
-void main() {
-  Future<int> testFuture(bool crash) async {
-    if (crash) {
-      throw Exception('error');
-    }
-    return 1;
+Future<int> testFuture(bool crash) async {
+  if (crash) {
+    throw Exception('error');
   }
+  return 1;
+}
 
+void main() {
   test('should return AsyncData when future is successful', () async {
     final result = await AsyncValue.tryCatch(() => testFuture(false));
     expect(result, isA<AsyncData>());
